bugfix - fixed error when submitting empty profile edit form

+++ app/routs.py
@@ -29,30 +29,30 @@ def login():
 def profile():
     form = EditProfileForm()
     if form.validate_on_submit():
-        f = form.picture.data
-        f.seek(0, 2)
-        size = f.tell()
-        f.seek(0)
-        if size > MAX_PP_SIZE:
-            flash(f'Soubor je příliš velký (max {int(MAX_PP_SIZE/(1024**2))} MB).', 'danger')
-            return redirect(url_for('profile'))
+        pp = form.picture.data
+        if pp:
+            pp.seek(0, 2)
+            size = pp.tell()
+            pp.seek(0)
+            if size > MAX_PP_SIZE:
+                flash(f'Soubor je příliš velký (max {int(MAX_PP_SIZE/(1024**2))} MB).', 'danger')
+                return redirect(url_for('profile'))
 
-        ext = os.path.splitext(secure_filename(f.filename))[1].lower()
-        filename = uuid.uuid4().hex + ext
-        pp_dir = os.path.join(app.instance_path, 'pp')
-        os.makedirs(pp_dir, exist_ok=True)
+            ext = os.path.splitext(secure_filename(pp.filename))[1].lower()
+            filename = uuid.uuid4().hex + ext
+            pp_dir = os.path.join(app.instance_path, 'pp')
+            os.makedirs(pp_dir, exist_ok=True)
 
-        if current_user.pp_filename:
-            old_path = os.path.join(pp_dir, current_user.pp_filename)
-            if os.path.exists(old_path):
-                os.remove(old_path)
+            if current_user.pp_filename:
+                old_path = os.path.join(pp_dir, current_user.pp_filename)
+                if os.path.exists(old_path):
+                    os.remove(old_path)
 
-        f.save(os.path.join(pp_dir, filename))
-        current_user.pp_filename = filename
-        db.session.commit()
-        flash('Profilová fotografie byla aktualizována.', 'success')
+            pp.save(os.path.join(pp_dir, filename))
+            current_user.pp_filename = filename
+            db.session.commit()
+            flash('Profilová fotografie byla aktualizována.', 'success')
         return redirect(url_for('profile'))
-
     return render_template('profile.html', form=form)
 
 @app.route('/uploads/pp/<filename>')
